macOS build.sh: Use default macosx SDK for xcodebuild targets
xcodebuild only allows to use an SDK which is bundled inside the
corresponding Xcode.app package. If you pass a different SDK
stored in another location, it bails out with:

xcodebuild: error: SDK "..." cannot be located.

Therefore, do not set an external SDK for contribs, but use
the default one.

This restores the behaviour we had previously for the 3.x branch.