qt: medialib: pass loader via a raw pointer
Currently, the loader is owned by MLModelCache, but in the future it
might outlive it: database queries will be executed from a separate
thread, and MLModelCache could be deleted while an asynchronous request
is running.

To prepare for this change, do not require to receive the loader via a
std::unique_ptr (let the cache wrap it).

Signed-off-by: Pierre Lamot <pierre@videolabs.io>