dvbsub: clarify a note
this note unhelpfully did not specify which version of the standard, nor
the section title, causing unnecessary difficulty identifying the
applicable text.

i identified the applicable section as being titled 'Object data segment'
which, in versions 1.5.1 and 1.6.1 is now located at section 7.2.5 rather
than 7.2.4 (from the official standards linked at [1]).

i have thus:
 - added a version number, along with section title.
 - updated the section number to go along witht hat version number.
 - switched to use the term 'section' rather than paragraph.

i haven't updated the rest of the references yet since for whatever reason
i neglected to do so when i originally made this patch and i'm too out of
touch with it now.

[1]: https://dvb.org/specifications/