httpd: do not disable timeout for RTSP
Use the same timeout for the RTSP connection as for the RTSP session.
A shorter timeout is inefficient, as it requires clients to reconnect
to refresh the session anyway. A longer timeout is pointless as the
session would timeout and the RTP streams would stop regardless of the
connection staying alive.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>