httpd: do not disable timeout for RTSP
Use the same timeout for the RTSP connection as for the RTSP session.
A shorter timeout is inefficient, as it requires clients to reconnect
to refresh the session anyway. A longer timeout is pointless as the
session would timeout and the RTP streams would stop regardless of the
connection staying alive.

Refs #25812.

(cherry picked from commit 38d214bc4f2ef68dfe6859383e12a20ae91e62de)