audiotrack: use AudioTrack.getLatency()
Instead of AudioSystem.getOutputLatency().

That way, we are sure the get the latency of the current output device.

Note: both methods are hidden and should not be used in favor or
AudioTrack.getTimestamp(). Unfortunately, getTimestamp() returns a valid
timestamp too late (after few seconds) or doesn't work with some phones.
Therefore, the usage of the hidden API is still needed as a backup plan.

cf. https://github.com/google/ExoPlayer/issues/5763

The android team allow the usage of this hidden API, waiting for a new
method in next Android versions.