qt: medialib: rename MLParentId to MLItemId
The item id sometimes represents the id of the "current" item, sometimes
the id of the "parent" item. Therefore, the fact that it is a "parent"
id depends on the context, and is not intrinsic to the type.

This caused confusion when getId() returned a MLParentId, representing
the id of the current item.

Rename to MLItemId to clarify.

Signed-off-by: Pierre Lamot <pierre@videolabs.io>