qt: medialib: define a super-type for list items
Concrete medialib item types are defined by the template parameter on
MLSlidingWindowModel.

Create a super-type to be able to "untemplatize" MLSlidingWindowModel
(in order to simplify and improve compilation times).

As a side-effect, this also makes explicit the necessity to define a
getId() method (called from MLSlidingWindowModel), a constraint which
could not be explicitly expressed on a template parameter.

Signed-off-by: Pierre Lamot <pierre@videolabs.io>