qt: medialib: un-templatize MLSlidingWindowModel
MLSlidingWindowModel was generic over its item type, and with many
non-template subclasses, each subclasses defining the actual concrete
type.

Therefore, this static genericity was "internal". In practice, it didn't
bring a lot a value, but it caused every subclass to recompile
the whole implementation.

To improve compilation times, remove the template parameter and always
use the item super-type MLItem. When an item of the concrete type is
necessary (once per subclass), just cast locally.

This change alone does not improve complitation times significantly,
because the implementation is still in the header file, which also
includes the whole ListCache implementation. The following commits will
solve these problems.

Signed-off-by: Pierre Lamot <pierre@videolabs.io>