android/audio: add AAudio
https://developer.android.com/ndk/guides/audio/aaudio/aaudio

AAudio is a native PCM-only audio API starting Android 8.0 (but enabled
in VLC starting Android 9.0 due to some bugs).

Having tested it with few device. The latency reported by this new API is
correct and without any irregularity, contrary to the
AudioTrack.getTimeStamp() API that is very broken.

Plus side, a huge gain of performance due to the absence of JNI C->JAVA->C++.

-40ms of latency reported by Sync-One2 versus -150ms with AudioTrack.

The remaining 40ms of latency to fix are likely on the mediacodec/vout side.