macosx: codesign: Check whether vlc-cache-gen can be executed
Check whether vlc-cache-gen can actually be executed, to
detect whether it is available for the right architecture.

If it is not available, support the possibility to not
sign the plugins.dat cache file. Also support the case
the cache file is not there from the beginning.

(manual bp of 18271e16182312cc2d0a80848d562e52efc2ebf8)