variables: privatise the callback CV
The condition variable was shared across all variables of a given
object to save space, but it ends up waking up the wrong waiting
threads.

vlc_cond_t is not as large as it used to be. In fact, it could be as
small as a single pointer: if we required holding the mutex during
signal/broadcast. The nested mutex is only there to handle unprotected
signalling.