video_output:control: merge the two wait conditions
The code using the is_held (outside the vout thread doing things on the vout)
is still blocked by the cond_timedwait but doesn't need an extra boolean for
that.

The wait for is_held is done earlier as it's more logical to wait for outside
calls to be finished blocking the vout thread loop before actually waiting for
a deadline, which may have passed anyway. In both cases (before and now), new
calls to control_Hold() were/are blocked until control_Pop() is done.

Moving this code also mimicks better the way control_Hold/control_Release are
using is_held and the (shared) wait condition. Once the timed wait is done we
are back to a (signaled) state similar to control_Release.