video_output:control: rename can_sleep to forced_awake
Invert the values as it means the opposite.
Explain why we keep forced_awake true while there are pending mouse events in
the FIFO.

This might also be handled out of control.c. The wake could just cause the
cond_timedwait to exit but the state to ignore the deadline and keep popping
mouse states from the control FIFO could be handled in video_output.c.