qml: medialib: Continue watching video section added to the section "Video"
"Continue watching" section with list of videos the user started playing
added at the top of the "Video" section.
Scrolling the "Video" page is performed on all area vertically.
Scrolling the "Continue watching" section is horizontal.
In order to preserve loading items on scrolling the content the Continue watching
video section inserted at the top as header of grid view and of list view.

Signed-off-by: Pierre Lamot <pierre@videolabs.io>