medialib: fix initialization
The medialib initialization should be called exactly once.

This was implemented by two mechanisms. Firstly, m_ml->isInitialized()
was checked to avoid any work if the initialization already occurred.
But since the medialib mutex was hold only during the call to
isInitialized(), as soon as it returns, another thread may have set
m_initialized. To handle this case harmlessly, Init() checks if
m_ml->initialize() returns AlreadyInitialized.

However, between the call to isInitialized() and initialize(), several
medialibrary methods must be called (registerDeviceLister() and
addFileSystemFactory()). These methods assert that m_initialized is
false, which was not guaranteed.

To fix this race condition, use our own flag protected by a local mutex,
hold during the whole Init() execution.

This also avoids a lot of unnecessary blocking calls: many medialibrary
controls (including the asynchronous ones) need to call Init() just to
make sure that the medialib was initialized. But since isInitialized()
is protected by the medialibray mutex, all these calls had to wait for
the mutex (hold during the execution of SQL queries) to be released.
With a separate mutex, this problem disappears.

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>