viewpoint: remove vlc_viewpoint_reverse
The function was doing the wrong operation. The underlying rotation,
which is R = Rz(roll) Rx(pitch) Ry(yaw), was transformed into the
product Rz(-roll) Rx(-pitch) Ry(-yaw), whereas the correct transformation
would have been Ry(-yaw) Rx(-pitch) Rz(-roll), which leads to a reversed
rotation order. In simpler term, this was a good idea to help understand
what was happening but it was wrongly executed.

Removing the function and using the correct input for writing the matrix
simplifies the code and makes the future quaternion code more logical.
The correct version of this process can be done later if still needed.