viewpoint: rotate 360 sphere and cube to -pi/2
Renderers in OpenGL and D3D11 were adding +pi/2 to yaw when generating
the View matrix. This additional +pi/2 is ought to be understood as a
-pi/2 angle on yaw instead since the matrix is written transposed.

This additional angle stems from the sphere coordinates being mapped
from 0 to 2.pi while the texture itself is conceived to be mapped on
the segment [-pi/2; 3.pi/2] instead.

This patch "hardcode" the -pi/2 rotation in the sphere mesh by matching
the generated coordinates with the ones expected for the texture
mapping, thus removing the pi/2 rotation in the View matrix.

The rotation was initially added in the following commit:
08207425d8d260a6650a332c4c7c7b040ff48b15

No additional rotation is done for D3D11 cube because it was already
accidentally rotated in ec0b151f397bf4470f58aa55daa9f7b2068113a4.