android: utils: add releaseTexImage
Add support for releaseTexImage in SurfaceTexture, which is needed in
custom SurfaceTexture-based buffer producer to use the less number of
buffer possible. This is especially important on old and low-memory
devices which cannot create a ton of SurfaceTexture objects.

Use the JNI variant for the NDK implementation since the function
doesn't even exists in the NDK.