android: utils: add get_texture to picture_context
Abstract the storage of SurfaceTexture in the Android video context, so
that the interop doesn't have to bother on how the SurfaceTexture is
handled by the image producer. In particular, this will allow multiple
SurfaceTexture to be used in a single Android context, thus provides the
ability to use SurfaceTexture as a single buffer in each picture_t sent
from an Android producer.