subtitle: fix AQT subtitle parsing
The numbers we read are frame numbers. There's a start and a stop frame number,
not just the start frame. The stop frame number was discarded so far as the
next start frame.
(based on http://web.archive.org/web/20070210095721/http://www.volny.cz/aberka/czech/aqt.html
and libavformat)