video_output: count the time to do the prepare to determine if a frame is late
The system_pts is the date we should do the swap (vd->display) for the picture.
But between now and the time we do the swap, there is also the prepare
(+ static filtering) that will happen. We are late if now+that duration is
after the date we should do the swap.