video_output: account for the time spent in static filters to check late frames
This time is not included in the sys->render counter which only counts the user
filters + prepare.
The static filters (deinterlacing) may take significant time and should be
checked to determine if a frame might be late or not.

By default we assume the time it takes is 0ms because by default deinterlacing
doesn't happen (on the first frame).