video_output: only get the next decoded picture when the current one is late
If displayed.current is bound to be displayed late, we look for the next
decoded (+statically filtered) picture that is not late.

We do not preemptively store the next picture in displayed.next anymore. If we
get a new picture we will overwrite displayed.current with it.

The date_refresh (new deadline) is computed once we know if we are using a new
picture or program a redisplay of the current one.