transcode: video: fix insertion of fps filter
The fps filter must be added when the target framerate is different
from the source framerate, not when we define the fps parameter.

We also check that fps.den > 0 to avoid divide-by-zero errors.

Probably regression from 030183d79045a1747d891572b2fe77d8cd0fd946 in
which the target framerate was defined in the destination format.