glspectrum: use symbols from vlc_gl_t instance
Use the symbols from the vlc_gl_t instance for OpenGL instead of linking
the platform OpenGL library.

This effectively decouples the glspectrum visualization from the
underlying OpenGL implementation, allowing alternative OpenGL
implementation or function wrapper to be used with it, and simplifying
the linkage of the plugin on every platforms.

Thanks to this decoupling, it's now possible to use the visualization
through OpenGL video callbacks in VLC.

Fixes #22520