apple: add support for partial linking
Each plugin built as archive library will have its object partially
linked before getting ar-chived. The work here is also partial since
dependencies are not partially-linked into the plugin too. In the
future, it will help prevent clashing of symbols between plugins and
reduce bloats in the final library archive on the current iOS static
build.

Note that this trades optimization window for less object bloating and
the guarantee that each plugin is only made of one object. Next step is
to be able to include the contribs and properly localize the symbols
that are not meant to be exposed for the VLC plugin ABI.