video_filter: Makefile.am: fix ci_filters's LDFLAGS
The libci_filter_plugin_la_LDFLAGS must always contains the value of
$(AM_LDFLAGS) since it is at least defined once. Otherwise, automake
won't automatically define it to $(AM_LDFLAGS) $(LDFLAGS) and the plugin
won't be built correctly (ie, built as a convenience library because of
the lack of -rpath definition in LDFLAGS).

In addition, the frameworks were not correctly set for iOS, which was
not a problem in static build because there is no linkage, nor in
dynamic build before because the plugin was built as a convenience
library, thus not linked nor installed to the device with the bundle
script.

Since the frameworks are mostly common between MacOSX and iOS, just add
the corresponding OpenGL framework for each target, and define the other
frameworks out of any automake condition.

Cherry-picked from commit c82c99522446be5f50296721d5716ae604eed426 but
partially modified to remove OpenGL/ES frameworks that are not needed
without the commit cb6b1d5fb6c28893479f737a94da2bc753cbe8b9.