VLCOpenGLES2VideoView: handle state change
Avoid making vlc_gl_MakeCurrent fail because the application would be
be moved into the background. Since the rendering is now disabled
directly by the vout_window, some parts of the observers on
NSNotificationCenter are now useless. What's left might or might not be
useless yet, because of the initial startup where the interaction with
the display is needed to create it anyway, so it's currently kept with a
condition variable to wait the application to become active instead of
bailing out.