vout: don't clobber crop borders when updating
The vout_UpdateSourceCrop() function uses the crop borders as both
inputs and outputs. There does not seem to be any good reason to
overwrite the values though.

Within vout_UpdateSourceCrop(), the value should actually not change
(much like the aspect ratio numerator and denominator in the previous
patchset), except for rounding errors and out-of-bound results.

The values are used in only one other function, vout_SetDisplayCrop().
There, old and new values are compared to avoid needlessly resetting
the display. Those comparisons actually work correctly only if the
values are *not* modified by vout_UpdateSourceCrop().