libvlc: the setup/cleanup rendering callbacks can be called many times
In some cases the decoder might ask for a decoder device that will use the
setup callback. But in the end it's not using the decoder device. Then the
display module is created, without knowing about that decoder device (via a
video context) and calls the setup callback again.

Unless the decoder device that is created in the vout is passed to the display
module, we need to allow those multiple calls per session.

Fix the D3D11 sample to take this in account. The D3D9 sample already takes
care of it.