VLCOpenGLES2VideoView: handle UIView resize
The UIView resize was previously handled using mostly UIView framework.
This is now forwarded by the vout_window through the vout_window_Report*
mechanism, under display_lock, mimicking the behaviour of every other
window system integration.

Because no feedback is currently forwarded to the window when the vout
thread has drawn something, it means that we cannot synchronize the
rendering with the end of the resize event from the main thread, so the
resize is out-of-sync.

Having the UIView being resized by the resize callback is necessary
though to move the generation of buffers out of the MakeCurrent callback
and making it not faillible.