d3d_dynamic_shader: simplify the range conversion handling
We can only convert to/from studio range once. The special case where the YUV
to RGB conversion is handled by the matrix can be handled before deciding the
range conversion to apply.

In the case of YUV full range the YUV->RGB will convert it to "extra full" so
we must scale back the range.