VLCOpenGLES2VideoView: use dedicated CFRunLoop mode
The vlc_runloop mode is designed to be executed even when an event is
being reported by the vout_window used, ie when there is a call to
CFRunLoopInMode(CFRunLoopGetMain(), CFSTR("vlc_runloop"), ..).

By adding the default mode too, it ensures it would run in the normal
CFRunLoop too.

Async tasks can still be dispatched without the "vlc_runloop" mode but
every sync task in the display must be done under this mode to prevent
deadlock from happening between the main thread and the vout_thread.

Fixes #23571