qt: medialib: reload asynchronously
On Qt view initialization completed, a medialibrary "reload" is
requested.

This call, which could take several seconds in practice (especially when
the medialibrary is initialized for the first time), was issued on
the UI thread, freezing the whole interface on starting.

Execute it from the Qt medialibrary thread pool instead.

Signed-off-by: Pierre Lamot <pierre@videolabs.io>