package: macosx: forward ac_cv symbols to $@
configure will store the command used when reconfiguring as long as the
ac_cv_ variables are passed as parameter to the configure execution
instead of being an environment variable.

So as to stay compatible with other clients of this script, use export
as default command when no argument is given to keep the previous
behaviour.