demux: ts: fix MPGA and ADTS probing
MPEG-1 Audio Layer II (MP2) was erroneously detected as ADTS.

MPGA: The previous check used wrong bitmasks (1 bit too far to the right).
The layer check was also wrong, the layer must not be 0.

ADTS: Previously only one bit of the layer was checked. Check both bits.
The check is now identical to the one in HasADTSHeader() in mpeg4audio.c.

References:
- http://mpgedit.org/mpgedit/mpeg_format/mpeghdr.htm
- https://wiki.multimedia.cx/index.php/ADTS

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>