contrib: nfs: Patch win32_compat.h to not override snprintf
Mingw toolchains always provide snprintf, and modern MSVC versions
do too.

The define of snprintf to sprintf_s is applied before stdio headers
are included, causing it to be applied to the headers themselves,
breaking mingw-w64 headers targeting UCRT (because the headers provide
declarations for both snprintf and sprintf_s, but one of them is
declared as a static inline while the other is not).

In upstream libnfs, this define was entirely commented out in
6564abaa1f979c0b62044f695030a32b60620163, so this backports one bit
of that commit.

(cherry picked from commit 970248c1e409b1fef0d8f50f97de6259c6f849a6)
Signed-off-by: Thomas Guillem <thomas@gllm.fr>