mmal: Makefile.am: fix mmal plugin definition
Use only the MMAL_CFLAGS/MMAL_LIBS instead of sharing the flags of a
non-existant plugin. In addition, LDFLAGS is for linker flags different
from the -l or -L ones, and -rpath only needs to be defined when using
conditional compilation from autoconf (LTLIB and VLC_ADD_PLUGIN) so the
MMAL_LIBS must go to LIBADD.

In prevision of the removal of the recursive Makefile.am, use the
MMAL_CFLAGS on every targets that needs them.