libvlc: let D3D11 apps provide a mutex to protect the ID3D11DeviceContext
ID3D11DeviceContext is a resource than should be protected in some cases. In
particular the hardware decoders must not do concurrent calls. So any use of
the same ID3D11DeviceContext must be protected as well. In VLC it is handled
but when the ID3D11DeviceContext comes from the host app it must also to the
same protection when using it outside of libvlc calls.

Setting this mutex in a GUID_CONTEXT_MUTEX private data is reserved for legacy
(UWP) apps.