opengl: make direct sampler support multi-texture
A "direct" sampler (i.e. a sampler without interop) only supported one
texture, because the output of the previous filter was assumed to be
RGBA in a single plane.

A deinterlace filter will output several planes, so the sampler of the
next filter must be able to receive all of them.

Signed-off-by: Alexandre Janniaux <ajanni@videolabs.io>