qt: qml: use enum for player controlbar identifier
Instead of using QString as the controlbar profile
model player identifier, use enum (int).

This patch constitutes the second step of the
integration of the new QML toolbar editor and the
refactored profile handling. Before this patch,
Players had to be referred again in the toolbar
editor which is bad for maintenance. Now, when
a player is added to the PlayerIdentifier enum,
Players should refer to the enum for their
identifiers.

Signed-off-by: Pierre Lamot <pierre@videolabs.io>