youtube.lua: disable cookies if redirected to consent page
In the past few days, YouTube has started redirecting requests for video
pages to a cookie consent and preference prompt, on a whole separate
page and domain; which prevents playback. We are not interested in
YouTube cookies, nor in consenting to them on behalf of our users, so
this just retries with cookies disabled. YouTube gets the hint and
simply redirects back to the original video page.

Fixes #25616

(cherry picked from commit f21c063ea3b8050541e5c8f4430378dfec85f4a5)
Signed-off-by: Pierre Ynard <linkfanel@yahoo.fr>