wgl: fix failing wglMakeCurrent
The context and original HDC is from the decoder thread. Close() is
called from the vout_thread when the vout is being stopped.

This leads to one of the following error when stopping the media player:

 - error 2004: The requested transformation operation is not supported.
 - error 6: The handle is invalid

Those errors means that either the hDC is still in use in another thread
or that the hDC is not valid anymore.

Since this was happening on Close(), there was no way to make it fail
gracefully.

Getting a new DC when making the context current, thus in the closing
thread too, fix the error by ensuring thread-safety with those objects.